Add workbooks POST API & username, email GET API, workbooks GET API
